refactor(app): slim App.tsx to 1 line by moving implementation to GameRoot.tsx. Lint+build green
